refac: remove INCLUDE: and add autocomplete module
