+++ playwright-tests/tests/communities.spec.js
@@ -1,4 +1,4 @@
-const { test, expect } = require("@playwright/test");
+import { test, expect } from "@playwright/test";
 
 test.describe("Wallet is connected", () => {
   test.use({
