+++ src/devhub/entity/community/Sidebar.jsx
@@ -41,7 +41,6 @@ return community === null ? (
       props={{
         fullWidth: true,
         minHeight: 0,
-        noBorder: true,
         children: <CommunitySummary />,
         style: { marginTop: "0.5rem" },
       }}
@@ -67,7 +66,6 @@ return community === null ? (
 
         fullWidth: true,
         minHeight: 0,
-        noBorder: true,
       }}
     />
   </div>
