+++ src/devhub/entity/post/PostEditor.jsx
@@ -1,33 +1,18 @@
-/* INCLUDE: "core/lib/autocomplete" */
-const autocompleteEnabled = true;
+const {
+  autoCompleteAccountId,
+  autocompleteEnabled,
+  AutoComplete,
+  textareaInputHandler,
+} = VM.require("${REPL_DEVHUB}/widget/core.lib.autocomplete");
 
-const AutoComplete = styled.div`
-  z-index: 5;
-
-  > div > div {
-    padding: calc(var(--padding) / 2);
-  }
-`;
-
-function textareaInputHandler(value) {
-  const showAccountAutocomplete = /@[\w][^\s]*$/.test(value);
-  State.update((lastKnownState) => ({
-    ...lastKnownState,
-    text: value,
-    showAccountAutocomplete,
-  }));
+if (
+  !autoCompleteAccountId ||
+  !autocompleteEnabled ||
+  !AutoComplete ||
+  !textareaInputHandler
+) {
+  return <p>Loading modules...</p>;
 }
-
-function autoCompleteAccountId(id) {
-  let description = state.description.replace(/[\s]{0,1}@[^\s]*$/, "");
-  description = `${description} @${id}`.trim() + " ";
-  State.update((lastKnownState) => ({
-    ...lastKnownState,
-    description,
-    showAccountAutocomplete: false,
-  }));
-}
-/* END_INCLUDE: "core/lib/autocomplete" */
 
 const postType = props.postType ?? "Sponsorship";
 const parentId = props.parentId ?? null;
